Refactoring packages & minor misc. changes
