+++ src/main/java/com/blockscore/models/Company.java
@@ -1,7 +1,5 @@
 package com.blockscore.models;
 
-import com.blockscore.common.CorporationType;
-import com.blockscore.common.ValidityStatus;
 import com.blockscore.models.base.BasicResponse;
 import com.blockscore.net.BlockscoreApiClient;
 import com.blockscore.net.BlockscoreRestAdapter;
@@ -278,7 +276,7 @@ public class Company extends BasicResponse {
    * @return true if valid.
    */
   public boolean isValid() {
-    return ValidityStatus.VALID.isEqualTo(status);
+    return status.equalsIgnoreCase("valid");
   }
 
   /**
