+++ src/main/java/com/blockscore/models/CompanyDetails.java
@@ -1,6 +1,5 @@
 package com.blockscore.models;
 
-import com.blockscore.common.MatchRating;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import org.jetbrains.annotations.NotNull;
