+++ src/main/java/com/blockscore/models/CorporationType.java
@@ -1,4 +1,4 @@
-package com.blockscore.common;
+package com.blockscore.models;
 
 import org.jetbrains.annotations.NotNull;
 
@@ -6,7 +6,7 @@ import org.jetbrains.annotations.NotNull;
  * Corporation type.
  */
 public enum CorporationType {
-  CORP("corporation"),
+  CORPORATION("corporation"),
   LLC("llc"),
   PARTNERSHIP("partnership"),
   SOLE_PROPRIETORSHIP("sp"),
@@ -16,16 +16,6 @@ public enum CorporationType {
 
   CorporationType(@NotNull final String value) {
     this.value = value;
-  }
-
-  /**
-   * Returns if this matches.
-   *
-   * @param value  the value to test
-   * @return whether the value is equivalent to the enum
-   */
-  public boolean isEqualTo(final String value) {
-    return this.value.equalsIgnoreCase(value);
   }
 
   /**
@@ -44,8 +34,8 @@ public enum CorporationType {
    * @return the CorporationType enum value
    */
   public static CorporationType toEnum(@NotNull final String value) {
-    if (value.equalsIgnoreCase(CORP.toString())) {
-      return CORP;
+    if (value.equalsIgnoreCase(CORPORATION.toString())) {
+      return CORPORATION;
     } else if (value.equalsIgnoreCase(LLC.toString())) {
       return LLC;
     } else if (value.equalsIgnoreCase(PARTNERSHIP.toString())) {
