+++ src/main/java/com/blockscore/models/PaginatedResult.java
@@ -1,4 +1,4 @@
-package com.blockscore.models.results;
+package com.blockscore.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.jetbrains.annotations.NotNull;
