+++ src/main/java/com/blockscore/models/PersonDetails.java
@@ -1,8 +1,5 @@
 package com.blockscore.models;
 
-import com.blockscore.common.MatchRating;
-import com.blockscore.common.ResultStrength;
-
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
