+++ src/main/java/com/blockscore/models/error/BlockscoreErrorType.java
@@ -1,4 +1,4 @@
-package com.blockscore.common;
+package com.blockscore.models.error;
 
 import org.jetbrains.annotations.NotNull;
