+++ src/main/java/com/blockscore/net/BlockscoreApiClient.java
@@ -4,7 +4,7 @@ import com.blockscore.common.Constants;
 import com.blockscore.models.Candidate;
 import com.blockscore.models.Company;
 import com.blockscore.models.Person;
-import com.blockscore.models.results.PaginatedResult;
+import com.blockscore.models.PaginatedResult;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonInclude;
