+++ src/main/java/com/blockscore/net/BlockscoreRestAdapter.java
@@ -5,9 +5,9 @@ import com.blockscore.models.Candidate;
 import com.blockscore.models.Company;
 import com.blockscore.models.Person;
 import com.blockscore.models.QuestionSet;
-import com.blockscore.models.results.PaginatedResult;
-import com.blockscore.models.results.WatchlistHit;
-import com.blockscore.models.results.WatchlistSearchResults;
+import com.blockscore.models.PaginatedResult;
+import com.blockscore.models.WatchlistHit;
+import com.blockscore.models.WatchlistSearchResults;
 
 import org.jetbrains.annotations.NotNull;
 import retrofit.http.Body;
